feat: add api/v1/recruitment/{id}/applicant/{id}/scores

+++ src/main/java/com/aivle08/big_project_api/repository/EvaluationScoreRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Collection;
 import java.util.List;
 
-public interface EvaluationScoreRepository extends JpaRepository<EvaluationScore, Long> {
+public interface EvaluationScoreRepository extends JpaRepository<EvaluationScore, Integer> {
     List<EvaluationScore> findByApplicantId(Long applicantId);
 }
